make UpnpDiscoveryResult a QObject in order to be able to have slots in it
move UpnpDiscoveryResult into its own file

make it a sub class of QObject in order to be able to have slots in it. They
will be used to manage time outs and to manage the validity of discovered
resources with timers. Should allow to fill in holes in SSDP support.